feat:Extract RequestWorkspace and Stateful Request Pane Entities.
