+++ src/app/postman_app.rs
@@ -11,15 +11,15 @@ use gpui::{
 };
 
 mod chrome;
-mod request_editor;
+mod request_workspace;
 
-use request_editor::{RequestEditor, RequestEditorEvent};
+use request_workspace::{RequestWorkspace, RequestWorkspaceEvent};
 
 /// Application composition root. Feature-specific controls and task lifetimes live in child
 /// entities; this type only wires the shell together.
 pub struct PostmanApp {
     view_model: Entity<WorkspaceViewModel>,
-    request_editor: Entity<RequestEditor>,
+    request_workspace: Entity<RequestWorkspace>,
     request_runner: Entity<RequestRunner>,
     history_list: Entity<HistoryList>,
     _subscriptions: Vec<Subscription>,
@@ -34,37 +34,37 @@ impl PostmanApp {
     /// Dependency-injected constructor used by app hosts and black-box UI tests that need to
     /// observe the ViewModel without mutating the View through a second command surface.
     pub fn with_view_model(view_model: Entity<WorkspaceViewModel>, cx: &mut Context<Self>) -> Self {
-        let request_editor = cx.new(|cx| RequestEditor::new(view_model.clone(), cx));
+        let request_workspace = cx.new(|cx| RequestWorkspace::new(view_model.clone(), cx));
         let request_runner = cx.new(|_| RequestRunner::new());
         let history_list = cx.new(|cx| HistoryList::new(view_model.clone(), cx));
         let subscriptions = vec![
-            cx.subscribe(&request_editor, Self::on_request_editor_event),
+            cx.subscribe(&request_workspace, Self::on_request_workspace_event),
             cx.subscribe(&history_list, Self::on_history_selected),
         ];
 
         Self {
             view_model,
-            request_editor,
+            request_workspace,
             request_runner,
             history_list,
             _subscriptions: subscriptions,
         }
     }
 
-    fn on_request_editor_event(
+    fn on_request_workspace_event(
         &mut self,
-        _editor: Entity<RequestEditor>,
-        event: &RequestEditorEvent,
+        _workspace: Entity<RequestWorkspace>,
+        event: &RequestWorkspaceEvent,
         cx: &mut Context<Self>,
     ) {
         match event {
-            RequestEditorEvent::Execute(pending) => {
+            RequestWorkspaceEvent::Execute(pending) => {
                 let pending = pending.clone();
                 let view_model = self.view_model.clone();
                 self.request_runner
                     .update(cx, |runner, cx| runner.execute(pending, view_model, cx));
             }
-            RequestEditorEvent::Abort(send_id) => {
+            RequestWorkspaceEvent::Abort(send_id) => {
                 self.request_runner
                     .update(cx, |runner, _| runner.abort(*send_id));
             }
@@ -78,12 +78,13 @@ impl PostmanApp {
         cx: &mut Context<Self>,
     ) {
         let HistoryListEvent::RequestSelected(request) = event;
-        self.request_editor
-            .update(cx, |editor, cx| editor.load_request(request, cx));
+        self.request_workspace
+            .update(cx, |workspace, cx| workspace.load_request(request, cx));
     }
 
     fn new_request(&mut self, cx: &mut Context<Self>) {
-        self.request_editor.update(cx, RequestEditor::new_request);
+        self.request_workspace
+            .update(cx, RequestWorkspace::new_request);
     }
 }
 
@@ -106,7 +107,7 @@ impl Render for PostmanApp {
                     .flex()
                     .child(self.render_left_rail(cx))
                     .child(self.history_list.clone())
-                    .child(self.request_editor.clone()),
+                    .child(self.request_workspace.clone()),
             )
     }
 }
