--- src/app/postman_app/request_editor/panes/mod.rs
@@ -1,5 +1,0 @@
-mod authorization;
-mod body;
-mod headers;
-mod params;
-mod scripts;
