--- src/app/postman_app/request_editor/panes/scripts.rs
@@ -1,50 +1,0 @@
-use super::super::RequestEditor;
-use crate::ui::{
-    components::body_input::BodyInput,
-    theme::{CODE_BG, CODE_TEXT, FONT_UI, MUTED},
-};
-use gpui::{
-    div, px, rgb, Entity, FontWeight, InteractiveElement, IntoElement, ParentElement, Styled,
-};
-
-impl RequestEditor {
-    pub(in crate::app::postman_app::request_editor) fn render_script_editor(
-        &self,
-        title: &'static str,
-        hint: &'static str,
-        input: Entity<BodyInput>,
-        selector: &'static str,
-    ) -> gpui::AnyElement {
-        div()
-            .flex_1()
-            .min_h_0()
-            .flex()
-            .flex_col()
-            .gap_2()
-            .p_3()
-            .bg(rgb(CODE_BG))
-            .child(
-                div()
-                    .flex()
-                    .items_center()
-                    .justify_between()
-                    .font_family(FONT_UI)
-                    .child(
-                        div()
-                            .text_size(px(12.0))
-                            .font_weight(FontWeight::BOLD)
-                            .text_color(rgb(CODE_TEXT))
-                            .child(title),
-                    )
-                    .child(div().text_size(px(11.0)).text_color(rgb(MUTED)).child(hint)),
-            )
-            .child(
-                div()
-                    .debug_selector(move || selector.into())
-                    .flex_1()
-                    .min_h_0()
-                    .child(input),
-            )
-            .into_any_element()
-    }
-}
